Added basic framework for creating the draw grid
Added a function that creates 16 divs inside of a container div via DOM
manipulation. This works by using a for loop to create the divs and
append them to the parent node of "container". From here, I can add
some CSS to make it look more like a grid and be well on my way towards
something resmebling an actual Etch-a-Sketch.