Added functionality for changing the canvas
Added functions as outlined in the project file. newArea() will take a
user input, make sure that is valid, then use that value as the new
number of squares per side in the canvas, update the area and percentage
used so that the flex basis can be updated to make sure that the new
divs take up the appropriate space. It then clears the canvas and runs
gridMaker() again to generate the new canvas. I also added a shake()
function, which keeps the canvas size and background the same but resets
the background colors of the individual divs, mimicking the behavior of
a real life Etch-a-Sketch.

I will now be adding buttons to trigger these functions. After that, I
will begin looking into how to do add some of the optional features(
random color mode, opacity) from the project instructions.