Added new features, assigned buttons
The buttons now run the appropriate functions. Had to make some changes
to the way that the shake function was working. Initially I was not able
to toggle the animation class used to animate the shaking of the div.
This resulted in the animation playing on every other click of the
button. I had to look into forcing a reflow and then adding and removing
the class instead. Next up is adding the rainbow mode, and then making
sure that this will work for screens of different sizes.