Added full rainbow mode functionality
Rainbow mode now actually works. I figured I would make it work by
keeping track of whether or not rainbow mode was on or off via a
boolean. I think this was the right approach but it did lead me to make
changes to how the draw function worked in order to respect that. I also
had to add some additional functions to handle generating random rgb
values. I am fairly satisfied with stopping here, but just out of
curiousity I would like to try to implement the opacity feature in the
bonus section of the project outline.